chore: move torso and qmix network files
